impl validate_sighash_all_anyonecanpay.
