+++ contracts/otx-lock/src/validate/mod.rs
@@ -1,0 +1,5 @@
+mod sighash_all_acp;
+mod sighash_single_acp;
+
+pub(crate) use sighash_all_acp::validate_sighash_all_anyonecanpay;
+pub(crate) use sighash_single_acp::validate_sighash_single_anyonecanpay;
